Split controllers into Skyscraper, Elevator, and Floor.
Done quite a lot here. Skyscraper is our main controller. The other two
are just used as api endpoints. Checkout mvc.js too for progress.
