+++ src/ElevatorSharp.Web/Controllers/ElevatorController.cs
@@ -13,18 +13,6 @@ namespace ElevatorSharp.Web.Controllers
 {
     public class ElevatorController : Controller
     {
-        #region UI Actions
-        public ActionResult Index()
-        {
-            var viewModel = new ElevatorIndexViewModel
-            {
-                Player = "Test Player",
-                Title = "Elevator Sharp"
-            };
-            return View("Original", viewModel);
-        }
-        #endregion
-
         #region Elevator Events
         /// <summary>
         /// Triggered when the elevator has completed all its tasks and is not doing anything.
@@ -92,67 +80,5 @@ namespace ElevatorSharp.Web.Controllers
             return Content(json, "application/json");
         }
         #endregion
-
-        #region Floor Events
-        /// <summary>
-        /// Triggered when someone has pressed the up button at a floor. 
-        /// Note that passengers will press the button again if they fail to enter an elevator.
-        /// Maybe tell an elevator to go to this floor?
-        /// </summary>
-        /// <param name="viewModel"></param>
-        /// <returns></returns>
-        public ContentResult UpButtonPressed(UpButtonPressedViewModel viewModel)
-        {
-            var json = JsonConvert.SerializeObject(viewModel);
-            return Content(json, "application/json");
-        }
-
-        /// <summary>
-        /// Triggered when someone has pressed the down button at a floor. 
-        /// Note that passengers will press the button again if they fail to enter an elevator.
-        /// Maybe tell an elevator to go to this floor?
-        /// </summary>
-        /// <param name="viewModel"></param>
-        /// <returns></returns>
-        public ContentResult DownButtonPressed(DownButtonPressedViewModel viewModel)
-        {
-            var json = JsonConvert.SerializeObject(viewModel);
-            return Content(json, "application/json");
-        }
-        #endregion
-
-        #region Helper Methods
-        private static void SaveSkyscraper(Skyscraper skyscraper)
-        {
-            var cache = MemoryCache.Default;
-            cache.Set("skyscraper", skyscraper, new CacheItemPolicy { AbsoluteExpiration = DateTimeOffset.Now.AddDays(1) });
-        }
-
-        private static Skyscraper LoadSkyscraper()
-        {
-            var cache = MemoryCache.Default;
-            if (cache.Contains("skyscraper"))
-            {
-                return (Skyscraper)cache.Get("skyscraper");
-            }
-            return null;
-        }
-
-        private static void SavePlayer(IPlayer player)
-        {
-            var cache = MemoryCache.Default;
-            cache.Set("player", player, new CacheItemPolicy { AbsoluteExpiration = DateTimeOffset.Now.AddDays(1) });
-        }
-
-        private static IPlayer LoadPlayer()
-        {
-            var cache = MemoryCache.Default;
-            if (cache.Contains("player"))
-            {
-                return (IPlayer)cache.Get("player");
-            }
-            return new TestPlayer();
-        }
-        #endregion  
     }
 }
