+++ src/ElevatorSharp.Web/Controllers/SkyscraperController.cs
@@ -1,0 +1,62 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.Caching;
+using System.Web;
+using System.Web.Mvc;
+using ElevatorSharp.Domain;
+using ElevatorSharp.Web.ViewModels;
+
+namespace ElevatorSharp.Web.Controllers
+{
+    public class SkyscraperController : Controller
+    {
+        public ActionResult Index()
+        {
+            var skyscraper = new Skyscraper();
+            var player = new TestPlayer(); // TODO: Load external dll from LS#ers
+            skyscraper.LoadPlayer(player); // This will call the Init method on Player and hook up events
+
+            var viewModel = new SkyscraperIndexViewModel
+            {
+                Player = "Test Player",
+                Title = "Elevator Sharp"
+            };
+            return View("Original", viewModel);
+        }
+
+        #region Helper Methods
+        private static void SaveSkyscraper(Skyscraper skyscraper)
+        {
+            var cache = MemoryCache.Default;
+            cache.Set("skyscraper", skyscraper, new CacheItemPolicy { AbsoluteExpiration = DateTimeOffset.Now.AddDays(1) });
+        }
+
+        private static Skyscraper LoadSkyscraper()
+        {
+            var cache = MemoryCache.Default;
+            if (cache.Contains("skyscraper"))
+            {
+                return (Skyscraper)cache.Get("skyscraper");
+            }
+            return null;
+        }
+
+        private static void SavePlayer(IPlayer player)
+        {
+            var cache = MemoryCache.Default;
+            cache.Set("player", player, new CacheItemPolicy { AbsoluteExpiration = DateTimeOffset.Now.AddDays(1) });
+        }
+
+        private static IPlayer LoadPlayer()
+        {
+            var cache = MemoryCache.Default;
+            if (cache.Contains("player"))
+            {
+                return (IPlayer)cache.Get("player");
+            }
+            return new TestPlayer();
+        }
+        #endregion  
+    }
+}
