+++ src/ElevatorSharp.Web/Views/Skyscraper/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model ElevatorSharp.Web.ViewModels.ElevatorIndexViewModel
+﻿@model ElevatorSharp.Web.ViewModels.SkyscraperIndexViewModel
 <div class="header">
     <div class="container">
         <div class="row">
