+++ src/ElevatorSharp.Web/Views/Skyscraper/Original.cshtml
@@ -1,4 +1,4 @@
-﻿@model ElevatorSharp.Web.ViewModels.ElevatorIndexViewModel
+﻿@model ElevatorSharp.Web.ViewModels.SkyscraperIndexViewModel
 <div class="container">
     <div class="header">
         <h1 class="left">Elevator Saga <em class="emphasis-color">The elevator programming game</em></h1>
